% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listtype.R
\name{depth}
\alias{depth}
\title{depth}
\usage{
depth(x, current = 0L)
}
\arguments{
\item{x}{An object}

\item{current}{Starting level for the nesting. Default is 0 and there
is no reason to change that. If you set \code{current} to a value \code{x} the result will
be equal to \code{depth(...) + x}.}
}
\value{
The maximum nesting depth found anywhere in the list.
The result will (obviously) always be 1 for atomic vectors. Returns \code{NA} for
objects that are not vectors.
}
\description{
Determines the (maximum) depth of nesting of vectors.
}
\examples{
depth(list(1, 2, list(3, "a"), 4, 5))
depth(expression(u, 2, u + 0:9))
}
