% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functiontype.R
\name{.isS3Method}
\alias{.isS3Method}
\title{.isS3Method}
\usage{
.isS3Method(name, env = parent.frame())
}
\arguments{
\item{name}{Function or name of function.}

\item{env}{Base environment in which to look for function definition.}
}
\value{
TRUE/FALSE
}
\description{
.isS3Method
}
\keyword{internal}
