% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functiontype.R
\name{.knownInternalGenericS3}
\alias{.knownInternalGenericS3}
\title{.knownInternalGenericS3}
\usage{
.knownInternalGenericS3()
}
\value{
a vector of function names
}
\description{
Provide a list with internal generics.
}
\keyword{internal}
