% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functiontype.R
\name{.knownMethodsS3}
\alias{.knownMethodsS3}
\title{.knownMethodsS3}
\usage{
.knownMethodsS3(nsName = NULL)
}
\arguments{
\item{nsName}{A string identifying a namespace}
}
\value{
A list of names known to be S3 methods in the given
namespace.
}
\description{
.knownMethodsS3
}
\keyword{internal}
