% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huh.R
\name{.paradigm}
\alias{.paradigm}
\title{Determine the OOP paradigm of an object.}
\usage{
.paradigm(x)
}
\arguments{
\item{x}{An object}
}
\value{
Is x an implicit, S3, S4, RC, or R6 object.
}
\description{
Determine the OOP paradigm of an object.
}
\keyword{internal}
