% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how.R
\name{how}
\alias{how}
\alias{how.default}
\alias{how.list}
\alias{how.array}
\alias{how.matrix}
\alias{how.data.frame}
\alias{how.complex}
\title{how}
\usage{
how(x)

\method{how}{default}(x)

\method{how}{list}(x)

\method{how}{array}(x)

\method{how}{matrix}(x)

\method{how}{data.frame}(x)

\method{how}{complex}(x)
}
\arguments{
\item{x}{An object}
}
\value{
An object of class \code{huh.how}.
}
\description{
Prints how an object can be subset
}
\section{Methods (by class)}{
\itemize{
\item \code{how(default)}: Default handler

\item \code{how(list)}: For lists

\item \code{how(array)}: For arrays

\item \code{how(matrix)}: For matrices

\item \code{how(data.frame)}: For data frames

\item \code{how(complex)}: Complex numbers

}}
\examples{
how(matrix(1:4, 2))
}
