% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huh.R
\name{huh}
\alias{huh}
\title{huh}
\usage{
huh(x)
}
\arguments{
\item{x}{An R object}
}
\value{
A list of class \code{huh} containing the specification of the object.
}
\description{
Creates a detailed summary of the nature of an object.
}
\examples{
huh(mean)
huh(1:3)
}
