% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listtype.R
\name{is.mixed}
\alias{is.mixed}
\title{is.mixed}
\usage{
is.mixed(x, f = typeof, recursive = TRUE)
}
\arguments{
\item{x}{a list}

\item{f}{a function used on each element. \code{typeof()} by default, but any other function
can be used.}

\item{recursive}{By default \code{is.mixed()} goes through the entire nestings of the list
(default: \code{TRUE}). If \code{FALSE} only the first level will be analysed.}
}
\value{
\code{TRUE} if all objects in the list \code{x} yield the same result \code{f(x)}. \code{NA} if \code{x} is
not a list.
}
\description{
is.mixed
}
\details{
\code{is.mixed} returning \code{TRUE} implies that \code{unlist()} will cause some an implicit
type coercion on some list elements.
}
\examples{
is.mixed(list(1, 2, 2L:3L, 4, 5))
is.mixed(list(1L, 2L, 2L:3L, 4L, 5L))
}
