% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listtype.R
\name{is.nested}
\alias{is.nested}
\title{is.nested}
\usage{
is.nested(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} is the object \code{x} is a list that contains elements of type \code{list}.
Otherwise \code{FALSE}.
}
\description{
is.nested
}
\examples{
is.nested(list(1, 2, 3, 4, 5)) # FALSE
is.nested(list(1, 2, 2:3, 4, 5)) # FALSE
is.nested(list(1:2, 2:3, 3:4, letters, LETTERS)) # FALSE
is.nested(list(1, 2, list(3, "a"), 4, 5)) # TRUE
is.nested(list(list(1), list(2), list(3), list(4), list(5))) # TRUE
is.nested(NULL) # FALSE
}
