% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.huh.how}
\alias{print.huh.how}
\title{print.huh.how}
\usage{
\method{print}{huh.how}(x, ...)
}
\arguments{
\item{x}{A \code{huh.how} object}

\item{...}{Further arguments (currently ignored)}
}
\value{
\code{invisible(x)}
}
\description{
print.huh.how
}
\examples{
print(how(complex(1, 2, 3)))
}
