% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.huh}
\alias{print.huh}
\title{print.huh}
\usage{
\method{print}{huh}(x, ...)
}
\arguments{
\item{x}{A \code{huh} object}

\item{...}{Additional arguments passed on to \code{format()}.}
}
\value{
Returns \code{x} invisibly
}
\description{
A formatted output of a \code{huh} object to the console.
}
\examples{
print(huh(1:3))
}
